inserted 1 row into artist table

+++ sql_queries.py
@@ -19,7 +19,7 @@ CREATE TABLE IF NOT EXISTS users (user_id INT PRIMARY KEY, first_name VARCHAR, l
 """)
 
 song_table_create = ("""
-CREATE TABLE IF NOT EXISTS songs (song_id VARCHAR PRIMARY KEY, title VARCHAR, artist_id VARCHAR, year INT, duration INT)
+CREATE TABLE IF NOT EXISTS songs (song_id VARCHAR PRIMARY KEY, title VARCHAR, artist_id VARCHAR, year INT, duration FLOAT)
 """)
 
 artist_table_create = ("""
